Recompra_local/pyproject.toml
[tool.poetry]
name = "recompra-local"
version = "0.1.0"
description = ""
authors = ["Juan David Duque <jduque1989@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.17.2"
webdriver-manager = "^4.0.1"
requests = "^2.31.0"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.0"
pyarrow = "^15.0.0"
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
